Update directory paths to have parentheses
