--- app/(pages)/staff/logger.tsx
@@ -1,8 +1,0 @@
-// pages/staff/logger.tsx
-import { withAuth } from '@/lib/withAuth';
-
-function LoggerPage() {
-  return <div>Logger Page for Staff (add service logs)</div>;
-}
-
-export default withAuth(LoggerPage, 'staff');
